Do not spread interface for one class to modules
