+++ lib/global_phone.rb
@@ -1,3 +1,4 @@
+require 'global_phone/utils'
 require 'global_phone/context'
 
 module GlobalPhone
