+++ lib/global_phone/database.rb
@@ -1,9 +1,8 @@
-require 'global_phone/parsing'
 require 'global_phone/region'
 
 module GlobalPhone
   class Database
-    include Parsing
+    attr_reader :regions
 
     def self.load_file(filename)
       load(File.read(filename))
@@ -14,11 +13,24 @@ module GlobalPhone
       new(JSON.parse(json))
     end
 
-    attr_reader :regions
-
     def initialize(record_data)
       @regions = record_data.map { |data| Region.new(data) }
       @territories_by_name = {}
+    end
+
+    def parse(string, territory_name)
+      string = Number.normalize(string)
+      territory = self.territory(territory_name)
+      raise ArgumentError, "unknown territory `#{territory_name}'" unless territory
+
+      if starts_with_plus?(string)
+        parse_international_string(string)
+      elsif string =~ territory.international_prefix
+        string = strip_international_prefix(territory, string)
+        parse_international_string(string)
+      else
+        territory.parse_national_string(string)
+      end
     end
 
     def region(country_code)
@@ -36,13 +48,44 @@ module GlobalPhone
       "#<#{self.class.name}>"
     end
 
-    protected
-      def regions_by_country_code
-        @regions_by_country_code ||= Hash[*regions.map { |r| [r.country_code, r] }.flatten]
-      end
+    private
 
-      def region_for_territory(name)
-        regions.find { |r| r.has_territory?(name) }
+    def parse_international_string(string)
+      string = Number.normalize(string)
+      string = strip_leading_plus(string) if starts_with_plus?(string)
+
+      if region = region_for_string(string)
+        region.parse_national_string(string)
       end
+    end
+
+    def starts_with_plus?(string)
+      string[0, 1] == "+"
+    end
+
+    def strip_leading_plus(string)
+      string[1..-1]
+    end
+
+    def strip_international_prefix(territory, string)
+      string.sub(territory.international_prefix, "")
+    end
+
+    def region_for_string(string)
+      candidates = country_code_candidates_for(string)
+      Utils.map_detect(candidates) { |country_code| region(country_code) }
+    end
+
+    def country_code_candidates_for(string)
+      (1..3).map { |length| string[0, length] }
+    end
+
+    def regions_by_country_code
+      @regions_by_country_code ||= Hash[*regions.map { |r| [r.country_code, r] }.flatten]
+    end
+
+    def region_for_territory(name)
+      regions.find { |r| r.has_territory?(name) }
+    end
   end
 end
